+++ CHANGELOG.md
@@ -7,7 +7,7 @@ This file contains the list of changes made to pyjoulescope_ui.
 
 ## 1.0.50
 
-2024 Feb 10 [in progress]
+2024 Feb 11 [in progress]
 
 * Dropped Python 3.9 support as static methods are not callable until 3.10.
   See https://docs.python.org/3/whatsnew/3.10.html
@@ -18,7 +18,9 @@ This file contains the list of changes made to pyjoulescope_ui.
 * Improved "developer" mode.
   * Renamed "debug" mode to "developer" mode.
   * Automatically close developer widgets on setting disable.
-  * Added Publish Spy Widget.
+  * Renamed Debug widget to Profile widget.
+  * Added Publish Spy widget.
+  * Added Log View widget.
 * Added WindowStaysOnTopHint to HelpHtmlMessageBox  #245
 * Modified Export all data to work when streaming  #246
 * Added zoom to dual markers in Waveform widget  #243